fix: refactor nondaemon info command
