+++ src/tracer/src/nondaemon_commands.rs
@@ -1,6 +1,3 @@
-use colored::Colorize;
-use console::Emoji;
-use std::fmt::Write;
 use std::process::Command;
 
 #[cfg(target_os = "linux")]
@@ -8,117 +5,14 @@ use crate::utils::get_kernel_version;
 
 use crate::client::config_manager::{Config, ConfigLoader, INTERCEPTOR_STDOUT_FILE};
 use crate::common::constants::{
-    FILE_CACHE_DIR, LOG_FILE, PID_FILE, REPO_NAME, REPO_OWNER, STDERR_FILE, STDOUT_FILE,
+    FILE_CACHE_DIR, PID_FILE, REPO_NAME, REPO_OWNER, STDERR_FILE, STDOUT_FILE,
 };
 use crate::daemon::client::DaemonClient;
+use crate::utils::info_formatter::InfoFormatter;
 use anyhow::{bail, Context, Result};
 use std::result::Result::Ok;
 use tokio::time::sleep;
 use tracing::debug;
-
-const STATUS_ACTIVE: Emoji<'_, '_> = Emoji("🟢 ", "🟢 ");
-const STATUS_INACTIVE: Emoji<'_, '_> = Emoji("🔴 ", "🔴 ");
-const STATUS_WARNING: Emoji<'_, '_> = Emoji("🟡 ", "🟡 ");
-const STATUS_INFO: Emoji<'_, '_> = Emoji("ℹ️ ", "ℹ️ ");
-
-struct InfoFormatter {
-    output: String,
-    width: usize,
-}
-
-impl InfoFormatter {
-    fn new(width: usize) -> Self {
-        Self {
-            output: String::new(),
-            width,
-        }
-    }
-
-    fn add_header(&mut self, title: &str) -> Result<()> {
-        writeln!(
-            &mut self.output,
-            "\n┌{:─^width$}┐",
-            format!(" {} ", title),
-            width = self.width - 2
-        )?;
-        Ok(())
-    }
-
-    fn add_footer(&mut self) -> Result<()> {
-        writeln!(
-            &mut self.output,
-            "└{:─^width$}┘",
-            "",
-            width = self.width - 2
-        )?;
-        Ok(())
-    }
-
-    fn add_section_header(&mut self, title: &str) -> Result<()> {
-        writeln!(
-            &mut self.output,
-            "├{:─^width$}┤",
-            format!(" {} ", title),
-            width = self.width - 2
-        )?;
-        Ok(())
-    }
-
-    fn add_field(&mut self, label: &str, value: &str, color: &str) -> Result<()> {
-        let colored_value = match color {
-            "green" => value.green(),
-            "yellow" => value.yellow(),
-            "cyan" => value.cyan(),
-            "magenta" => value.magenta(),
-            "blue" => value.blue(),
-            "red" => value.red(),
-            "bold" => value.bold(),
-            _ => value.normal(),
-        };
-
-        // Calculate available space for value
-        let label_width = 20;
-        let padding = 4;
-        let max_value_width = self.width - label_width - padding;
-
-        // Format the value with proper truncation
-        let formatted_value = if colored_value.len() > max_value_width {
-            format!("{}...", &colored_value[..max_value_width - 3])
-        } else {
-            colored_value.to_string()
-        };
-
-        writeln!(
-            &mut self.output,
-            "│ {:<label_width$} │ {}  ",
-            label, formatted_value
-        )?;
-        Ok(())
-    }
-
-    fn add_status_field(&mut self, label: &str, value: &str, status: &str) -> Result<()> {
-        let (emoji, color) = match status {
-            "active" => (STATUS_ACTIVE, "green"),
-            "inactive" => (STATUS_INACTIVE, "red"),
-            "warning" => (STATUS_WARNING, "yellow"),
-            _ => (STATUS_INFO, "blue"),
-        };
-
-        writeln!(
-            &mut self.output,
-            "│ {:<20} │ {} {}  ",
-            label,
-            emoji,
-            value.color(color)
-        )?;
-        Ok(())
-    }
-
-    fn add_empty_line(&mut self) -> Result<()> {
-        writeln!(&mut self.output, "│{:width$}│", "", width = self.width - 2)?;
-        Ok(())
-    }
-}
 
 pub fn clean_up_after_daemon() -> Result<()> {
     std::fs::remove_file(PID_FILE).context("Failed to remove pid file")?;
@@ -284,8 +178,8 @@ pub async fn print_config_info(api_client: &DaemonClient, config: &Config) -> Re
         Ok(info) => info,
         Err(e) => {
             tracing::error!("Error getting info response: {e}");
-            print_error_state(&mut formatter)?;
-            println!("{}", formatter.output);
+            formatter.print_error_state()?;
+            println!("{}", formatter.get_output());
             return Ok(());
         }
     };
@@ -293,115 +187,15 @@ pub async fn print_config_info(api_client: &DaemonClient, config: &Config) -> Re
     formatter.add_header("TRACER INFO")?;
     formatter.add_empty_line()?;
 
-    print_daemon_status(&mut formatter)?;
+    formatter.print_daemon_status()?;
 
     if let Some(inner) = &info.inner {
-        print_pipeline_info(&mut formatter, inner, &info)?;
+        formatter.print_pipeline_info(inner, &info)?;
     }
 
-    print_config_and_logs(&mut formatter, config)?;
+    formatter.print_config_and_logs(config)?;
     formatter.add_footer()?;
-    println!("{}", formatter.output);
-    Ok(())
-}
-
-fn print_error_state(formatter: &mut InfoFormatter) -> Result<()> {
-    formatter.add_header("TRACER CLI STATUS")?;
-    formatter.add_empty_line()?;
-    formatter.add_status_field("Daemon Status", "Not Started", "inactive")?;
-    formatter.add_field("Version", env!("CARGO_PKG_VERSION"), "bold")?;
-    formatter.add_empty_line()?;
-    formatter.add_section_header("NEXT STEPS")?;
-    formatter.add_empty_line()?;
-    formatter.add_field("Interactive Setup", "tracer init", "bold")?;
-    formatter.add_field("Visualize Data", "https://sandbox.tracer.app", "blue")?;
-    formatter.add_field(
-        "Documentation",
-        "https://github.com/Tracer-Cloud/tracer-client",
-        "blue",
-    )?;
-    formatter.add_field("Support", "support@tracer.cloud", "blue")?;
-    formatter.add_empty_line()?;
-    formatter.add_footer()?;
-    Ok(())
-}
-
-fn print_daemon_status(formatter: &mut InfoFormatter) -> Result<()> {
-    formatter.add_section_header("DAEMON STATUS")?;
-    formatter.add_empty_line()?;
-    formatter.add_status_field("Status", "Running", "active")?;
-    formatter.add_field("Version", env!("CARGO_PKG_VERSION"), "bold")?;
-    formatter.add_empty_line()?;
-    Ok(())
-}
-
-fn print_pipeline_info(
-    formatter: &mut InfoFormatter,
-    inner: &crate::daemon::structs::InnerInfoResponse,
-    info: &crate::daemon::structs::InfoResponse,
-) -> Result<()> {
-    formatter.add_section_header("RUN DETAILS")?;
-    formatter.add_empty_line()?;
-
-    // Pipeline section
-    formatter.add_field("Pipeline Name", &inner.pipeline_name, "bold")?;
-    formatter.add_field(
-        "Pipeline Type",
-        inner.tags.pipeline_type.as_deref().unwrap_or("Not Set"),
-        "cyan",
-    )?;
-    formatter.add_field(
-        "Environment",
-        inner.tags.environment.as_deref().unwrap_or("Not Set"),
-        "yellow",
-    )?;
-    formatter.add_field(
-        "User",
-        inner.tags.user_operator.as_deref().unwrap_or("Not Set"),
-        "magenta",
-    )?;
-
-    // Run section
-    formatter.add_field("Run Name", &inner.run_name, "bold")?;
-    formatter.add_field("Run ID", &inner.run_id, "cyan")?;
-    formatter.add_field("Runtime", &inner.formatted_runtime(), "yellow")?;
-    formatter.add_field(
-        "Monitored Processes",
-        &format!(
-            "{}: {}",
-            info.watched_processes_count.to_string().bold(),
-            info.watched_processes_preview().cyan()
-        ),
-        "normal",
-    )?;
-    formatter.add_empty_line()?;
-    Ok(())
-}
-
-fn print_config_and_logs(formatter: &mut InfoFormatter, config: &Config) -> Result<()> {
-    formatter.add_section_header("CONFIGURATION & LOGS")?;
-    formatter.add_empty_line()?;
-
-    let grafana_url = if config.grafana_workspace_url.is_empty() {
-        "Not configured".to_string()
-    } else {
-        config.grafana_workspace_url.clone()
-    };
-
-    formatter.add_field("Grafana Workspace", &format!("{} ", grafana_url), "blue")?;
-    formatter.add_field(
-        "Process Polling",
-        &format!("{} ms ", config.process_polling_interval_ms),
-        "yellow",
-    )?;
-    formatter.add_field(
-        "Batch Submission",
-        &format!("{} ms ", config.batch_submission_interval_ms),
-        "yellow",
-    )?;
-    formatter.add_field("Standard Output", &format!("{} ", STDOUT_FILE), "cyan")?;
-    formatter.add_field("Standard Error", &format!("{} ", STDERR_FILE), "cyan")?;
-    formatter.add_field("Log File", &format!("{} ", LOG_FILE), "cyan")?;
+    println!("{}", formatter.get_output());
     Ok(())
 }
