+++ src/tracer/src/utils/mod.rs
@@ -1,3 +1,5 @@
+pub mod info_formatter;
+
 use crate::client::config_manager::TRACER_ANALYTICS_ENDPOINT;
 use crate::common::types::analytics::{AnalyticsEventType, AnalyticsPayload};
 use anyhow::Context;
